Refaktorert test m.m
