+++ core/src/main/java/org/gruppe3/core/service/TicketService.java
@@ -1,7 +1,12 @@
 package org.gruppe3.core.service;
 
+import java.util.ArrayList;
+
 import org.gruppe3.core.domain.Ticket;
 import org.gruppe3.core.dto.CreateTicketRequest;
+import org.gruppe3.core.dto.GetUserTicketsRequest;
+import org.gruppe3.core.dto.GetUserTicketsResult;
+import org.gruppe3.core.dto.TicketDTO;
 import org.gruppe3.core.exception.TicketRepositoryException;
 import org.gruppe3.core.port.TicketRepositoryPort;
 
@@ -17,5 +22,23 @@ public class TicketService {
     Ticket ticket =
         new Ticket(request.getTicketHash(), request.getTicketType(), request.getTicketRoute());
     ticketRepository.createTicket(ticket);
+  }
+
+  public GetUserTicketsResult getUserTickets(GetUserTicketsRequest request)
+      throws TicketRepositoryException {
+
+    ArrayList<Ticket> userTickets = ticketRepository.getUserTickets(request.getUserId());
+
+    ArrayList<TicketDTO> userTicketsResult = new ArrayList<>();
+
+    for (Ticket ticket : userTickets) {
+      TicketDTO ticketDTO =
+          new TicketDTO(ticket.getTicketId(), ticket.getTicketType(), ticket.getTicketRoute());
+      userTicketsResult.add(ticketDTO);
+    }
+
+    GetUserTicketsResult result = new GetUserTicketsResult(request.getUserId(), userTicketsResult);
+
+    return result;
   }
 }
